feat(app): Fixed some visual styling
